Front end almost done!

+++ app/src/main/java/org/giorgi/personalbudget/model/Transaction.java
@@ -7,17 +7,17 @@ package org.giorgi.personalbudget.model;
 public class Transaction {
     public static final String DATE_FORMAT = "dd/MM/yyyy hh:mm:ss";
 
-    private double amount;
+    private float amount;
     private boolean income;
     private String time;
 
-    public Transaction(double amount, boolean income, String time) {
+    public Transaction(float amount, boolean income, String time) {
         this.amount = amount;
         this.income = income;
         this.time = time;
     }
 
-    public double getAmount() {
+    public float getAmount() {
         return amount;
     }
